Moved scripts functionality into its own module.
